Initial commit. Currently does nothing but respond to status reqs.
With this, the initial part of a system for handling the Minecraft
protocol has been laid down, which I hope will be effective enough
to last into the beta of this software. Right now, only what's
needed to respond to a status request is included, though, which
is a Varint, String, and basic numerical types.

My code is a bit of a mess, with todo notes and poor code
everywhere, but please do forgive me, as this is my first commit
of the first Rust project I've ever written, and I didn't even
follow any tutorial to learn.

The license is GPLv3+, with a header in each file. Gitignore
includes exclusions for Rust and Emacs. Cargo produces a simple
bin, requiring the byteorder library. There is a placeholder
README.

Endless thanks to wiki.vg too, for all of you documenting the
Minecraft protocol so well.

That's all for the initial commit. Future commit messages will
be shorter, as they are incremental, instead of just being a huge
paste like this.